+++ lib/locale_switcher.dart
@@ -6,5 +6,6 @@ export 'package:locale_switcher/src/lang_icon_with_tool_tip.dart';
 export 'package:locale_switcher/src/locale_manager.dart';
 export 'package:locale_switcher/src/locale_switcher.dart';
 export 'package:locale_switcher/src/select_locale_button.dart';
+export 'package:locale_switcher/src/get_icon_for_language.dart';
 
 // export 'package:locale_switcher/src/locale_store.dart';