+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.7.0
+
+* provides a [LocaleManager.realLocaleNotifier] to dynamically change locale (and [LocaleManager.localeNotifier] to listen to locale changes),
+* factory constructors for [LocaleSwitcher]: [LocaleSwitcher.toggle], [LocaleSwitcher.menu] or [LocaleSwitcher.custom],
+* various...
+
 ## 0.6.4
 
 * added a lot of parameters,